made changes for percentage

+++ src/main/java/com/engati/data/analytics/engine/model/response/DashboardChartResponse.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @Data
 public class DashboardChartResponse {
 
-  private Map<String, Long> metricPercentage;
+  private Map<String, Double> metricPercentage;
 
 }
